render adjust

+++ src/Graphic/Mode1.h
@@ -7,7 +7,7 @@ class Mode1 :public Plain
 public:
 	Mode1(const Comment &comment);
 	QList<QRectF> locate();
-	bool move(qint64 time);
+	bool move(double time);
 	uint intersects(Graphic *other);
 
 private:
