+++ src/Graphic/Mode6.h
@@ -7,7 +7,7 @@ class Mode6 :public Plain
 public:
 	Mode6(const Comment &comment);
 	QList<QRectF> locate();
-	bool move(qint64 time);
+	bool move(double time);
 	uint intersects(Graphic *other);
 
 private:
